Rename gateway-tests to gateway
