+++ CODEOWNERS
@@ -367,7 +367,7 @@
 /tests/event-bus/ @montaro @abbi-gaurav @marcobebway @radufa @sayanh
 
 # Event Service Tests
-/tests/gateway-tests/ @montaro @abbi-gaurav @marcobebway @radufa @sayanh
+/tests/gateway/ @montaro @abbi-gaurav @marcobebway @radufa @sayanh
 
 # E2E Backup Tests
 /tests/end-to-end/backup-restore-test @a-thaler @igor-karpukhin @hisarbalik @lilitgh @suleymanakbas91 @clebs
