+++ tests/gateway/entrypoint.sh
